Bid draft is being saved locally
        Tender object is being obtained from BC and then the
        questionnaire is generated for the bid form to be submited.

        Reported By: Mesias Rendon <amrendonsa@unal.edu.co>

+++ src/renderer/components/common/form/Question.vue
@@ -228,7 +228,7 @@ export default {
           fileName = file.fileName;
         }
       });
-      if (fs.existsSync(path.join(this.localAnswer, fileName))) {
+      if (fileName && fs.existsSync(path.join(this.localAnswer, fileName))) {
         fs.unlink(path.join(this.localAnswer, fileName), (err) => {
           if (err) throw err;
         });
